Avoid conflicts build-essential
The `ensure_packages('build-essential', ...)` was conflicting with
another usage of `ensure_packages('build-essential')` elsewhere in our
puppet repository. This change declares the desired dependency in a way
that does not require setting parameters on the package resource.
Because the parameters are removed, the ensure_packages becomes more
generic and less likely to conflict with other modules that might define
the same package resource.